Moved SectorModel into its own module
